dags finales, logger y archivos de config y generacion de dags dinamicos
Para corregir el error de que estuve comiteando a otra rama de otro arbol de otro bosque...
